refactor(theme-classic): always collocate stylesheets with components in one folder
