+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/Sidebar/styles.module.css
