+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/index.tsx
@@ -11,10 +11,9 @@ import BackToTopButton from '@theme/BackToTopButton';
 import type {Props} from '@theme/DocPage/Layout';
 import DocPageLayoutSidebar from '@theme/DocPage/Layout/Sidebar';
 import DocPageLayoutMain from '@theme/DocPage/Layout/Main';
-
-import styles from './index.module.css';
-
 import {useDocsSidebar} from '@docusaurus/theme-common';
+
+import styles from './styles.module.css';
 
 export default function DocPageLayout({children}: Props): JSX.Element {
   const sidebar = useDocsSidebar();
