+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/styles.module.css
