+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Category/index.tsx
