+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Html/styles.module.css
