+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Link/index.tsx
@@ -14,7 +14,7 @@ import IconExternalLink from '@theme/IconExternalLink';
 
 import type {Props} from '@theme/DocSidebarItem/Link';
 
-import styles from './Link.module.css';
+import styles from './styles.module.css';
 
 export default function DocSidebarItemLink({
   item,
