+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Link/styles.module.css
