+++ packages/docusaurus-theme-classic/src/theme/MDXComponents/Img/index.tsx
@@ -7,8 +7,9 @@
 
 import React from 'react';
 import type {Props} from '@theme/MDXComponents/Img';
-import styles from './Img.module.css';
 import clsx from 'clsx';
+
+import styles from './styles.module.css';
 
 function transformImgClassName(className?: string): string {
   return clsx(className, styles.img);
