+++ packages/docusaurus-theme-classic/src/theme/MDXComponents/Img/styles.module.css
