+++ packages/docusaurus-theme-classic/src/theme/TOCCollapsible/CollapseButton/styles.module.css
