+++ packages/docusaurus-theme-classic/src/theme/TOCCollapsible/index.tsx
@@ -8,10 +8,11 @@
 import React from 'react';
 import clsx from 'clsx';
 import {useCollapsible, Collapsible} from '@docusaurus/theme-common';
-import styles from './index.module.css';
 import TOCItems from '@theme/TOCItems';
 import type {Props} from '@theme/TOCCollapsible';
 import CollapseButton from '@theme/TOCCollapsible/CollapseButton';
+
+import styles from './styles.module.css';
 
 export default function TOCCollapsible({
   toc,
