+++ packages/docusaurus-theme-classic/src/theme/TOCCollapsible/styles.module.css
